functions.py
"""
Define the function declaration and implementation for set_timecodes.
"""
from google.genai import types

# Example manual declaration (optional):
# set_timecodes_decl = types.FunctionDeclaration(
#     name="set_timecodes",
#     description="Set the timecodes for the video with associated text",
#     parameters=types.Schema(
#         type="OBJECT",
#         properties={
#             "timecodes": types.Schema(
#                 type="ARRAY",
#                 items=types.Schema(
#                     type="OBJECT",
#                     properties={
#                         "time": types.Schema(type="STRING"),
#                         "text": types.Schema(type="STRING"),
#                     },
#                     required=["time", "text"],
#                 ),
#             ),
#         },
#         required=["timecodes"],
#     ),
# )

def set_timecodes(timecodes: list[dict]) -> list[dict]:
    """
    Handler for 'set_timecodes'. Returns the provided timecodes.
    """
    return timecodes
